moved nexus templates to be part of future admin templates
