+++ experiments/templates/experiments/dashboard.html
